requirements.txt
anyio
async-timeout
certifi
charset-normalizer
click
dnspython
email-validator
fastapi
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
MarkupSafe
orjson
pydantic
Pygments
python-dateutil
python-dotenv
python-multipart
PyYAML
redis
requests
rfc3986
six
sniffio
starlette
typing_extensions
ujson
urequests
urllib3
uvicorn
uvloop
websockets
